+++ js/views/view/index.js
@@ -42,7 +42,7 @@ async function onApp(app) {
     }
     
     const { preferences, ListingDB } = app;
-    const limit = preferences.search_results_count || 1000;
+    const limit = preferences.settings.search_results_count || 1000;
     const table = ListingDB.listings;
     const collection = table.orderBy('index').reverse();
     const records = await table.orderBy('index').reverse().limit(limit).toArray();